+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to
 
 ### Changed
 
+- Implemented safeguards to prevent deletion of jobs with associated run history
+  [#1570](https://github.com/OpenFn/Lightning/issues/1570)
+
 ### Fixed
 
 ## [v2.0.0-rc9] 2024-02-05
@@ -92,7 +95,8 @@ and this project adheres to
 
 ### Fixed
 
-- Fix Run via Docker [#1653](https://github.com/OpenFn/Lightning/issues/1653)
+- Fix Run via Docker  
+  [#1653](https://github.com/OpenFn/Lightning/issues/1653)
 - Fix remaining warnings, enable "warnings as errors"
   [#1642](https://github.com/OpenFn/Lightning/issues/1642)
 - Fix workflow dashboard bug when viewed for newly created workflows with only